Typescript update: models and daos

+++ daos/passwordToken.ts
@@ -1,12 +1,10 @@
-const PasswordToken = require("../models/passwordToken");
-const User = require("../models/user");
-const crypto = require("crypto");
-const bcrypt = require("bcrypt");
-const sendEmail = require("../utils/email/sendEmail");
+import { hash as _hash, compare } from "bcrypt";
+import { randomBytes } from "crypto";
+import PasswordToken from "../models/passwordToken";
+import User from "../models/user";
+import sendEmail from "../utils/email/sendEmail";
 
-module.exports = {};
-
-module.exports.makePasswordTokenForUserEmail = async (email) => {
+export async function makePasswordTokenForUserEmail(email: string) {
   const bcryptSalt = Number(process.env.BCRYPT_SALT);
   const clientURL = process.env.CLIENT_URL;
   const user = await User.findOne({ email });
@@ -14,8 +12,8 @@ module.exports.makePasswordTokenForUserEmail = async (email) => {
   if (!user) throw new Error("Email not registered");
   let token = await PasswordToken.findOne({ userId: user._id });
   if (token) await token.deleteOne();
-  let resetToken = crypto.randomBytes(32).toString("hex");
-  const hash = await bcrypt.hash(resetToken, Number(bcryptSalt));
+  let resetToken = randomBytes(32).toString("hex");
+  const hash = await _hash(resetToken, Number(bcryptSalt));
 
   await new PasswordToken({
     userId: user._id,
@@ -32,7 +30,7 @@ module.exports.makePasswordTokenForUserEmail = async (email) => {
   );
 
   return { message: `Password reset email sent to ${user.email}` };
-};
+}
 
 // module.exports.findTokenByUserId = async (userId) => {
 //   return await Token.findOne({ userId });
@@ -47,33 +45,41 @@ module.exports.makePasswordTokenForUserEmail = async (email) => {
 //   return await Token.findOneAndDelete({ token: tokenString });
 // };
 
-module.exports.resetPassword = async (userId, token, password) => {
+export async function resetPassword(
+  userId: string,
+  token: string,
+  password: string
+) {
   const bcryptSalt = Number(process.env.BCRYPT_SALT);
 
   let passwordResetToken = await PasswordToken.findOne({ userId });
   if (!passwordResetToken) {
     throw new Error("Invalid or expired password reset token");
   }
-  const isValid = await bcrypt.compare(token, passwordResetToken.token);
+  const isValid = await compare(token, passwordResetToken.token);
   if (!isValid) {
     throw new Error("Invalid or expired password reset token");
   }
-  const hash = await bcrypt.hash(password, Number(bcryptSalt));
+  const hash = await _hash(password, Number(bcryptSalt));
   await User.updateOne(
     { _id: userId },
     { $set: { password: hash } },
     { new: true }
   );
   const user = await User.findById({ _id: userId });
-  sendEmail(
-    user.email,
-    "Password Reset Successfully",
-    {
-      name: user.username
-    },
-    "./template/resetPassword.handlebars",
-    "./template/resetPassword.txt.handlebars"
-  );
-  await passwordResetToken.deleteOne();
-  return { message: `Password reset was successful for ${user.email}` };
-};
+  if (user) {
+    sendEmail(
+      user.email,
+      "Password Reset Successfully",
+      {
+        name: user.username
+      },
+      "./template/resetPassword.handlebars",
+      "./template/resetPassword.txt.handlebars"
+    );
+    await passwordResetToken.deleteOne();
+    return { message: `Password reset was successful for ${user.email}` };
+  } else {
+    throw new Error("User not found");
+  }
+}
